Delet user completed with POST method. Yet to refine to accept DELETE method

+++ quora-api/src/main/java/com/upgrad/quora/api/controller/AdminController.java
@@ -3,6 +3,7 @@ package com.upgrad.quora.api.controller;
 
 import com.upgrad.quora.api.model.UserDeleteResponse;
 import com.upgrad.quora.service.business.AdminService;
+import com.upgrad.quora.service.business.SignoutService;
 import com.upgrad.quora.service.entity.UserEntity;
 import com.upgrad.quora.service.exception.AuthorizationFailedException;
 import com.upgrad.quora.service.exception.UserNotFoundException;
@@ -12,9 +13,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import com.upgrad.quora.api.model;
-
-
 @RestController
 @RequestMapping("/")
 public class AdminController {
@@ -22,13 +20,25 @@ public class AdminController {
     @Autowired
     private AdminService adminService;
 
+    @Autowired
+    private SignoutService signoutService;
 
-    @RequestMapping(method = RequestMethod.DELETE, path = "/admin/user/{userId}", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public ResponseEntity<UserDeleteResponse> deleteUser(@PathVariable("user_id") final String uuid, @RequestHeader("authorization") final String authorization) throws AuthorizationFailedException, UserNotFoundException {
 
-        UserEntity userEntity = adminService.deleteUser(uuid, authorization);
-        final UserDeleteResponse userDeleteResponse = new UserDeleteResponse().id(userEntity.getUuid()).status("USER SUCCESSFULLY DELETED");
-        return new ResponseEntity<UserDeleteResponse>(userDeleteResponse, HttpStatus.OK);
+    @RequestMapping(method = RequestMethod.POST, path = "/admin/user/{userId}", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<UserDeleteResponse> deleteUser(@PathVariable("userId") final String uuid, @RequestHeader("authorization") final String authorization) throws AuthorizationFailedException, UserNotFoundException {
+//        System.out.println("TEST");
+        if(signoutService.hasUserSignedIn(authorization)) {
+            if(signoutService.isUserAccessTokenValid(authorization)) {
+                String UUID = adminService.deleteUser(uuid, authorization);
+                //   System.out.println(userEntity.getUuid());
+                final UserDeleteResponse userDeleteResponse = new UserDeleteResponse().id(UUID).status("USER SUCCESSFULLY DELETED");
+                return new ResponseEntity<UserDeleteResponse>(userDeleteResponse, HttpStatus.OK);
+            } else {
+                throw new AuthorizationFailedException("ATHR-002", "User is signed out");
+            }
+        } else {
+            throw new AuthorizationFailedException("ATHR-001", "User has not signed in");
+        }
 
     }
 }
